masha: fix atom port not connected by host check
This commit checks for the validity of the atom_port
pointer in Masha, as hosts are known to instantiate the plugin
an not connect the Atom port, even though it is *not* marked as
connectionOptional.

This check avoids hosts crashing on running a plugin when the
atom port is NULL, and instead disables the Atom BPM functionality.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>